feat: rbac role res perm flag

+++ core/dao/model/rbac.go
@@ -127,9 +127,9 @@ type PermRoleRes struct {
 	// 编号
 	Id int64 `db:"id" pk:"yes" auto:"yes"`
 	// 菜单ID
-	ResId int64 `db:"res_id"`
+	ResId int `db:"res_id"`
 	// 角色ID
-	RoleId int64 `db:"role_id"`
+	RoleId int `db:"role_id"`
 	// 权限位值, 1:增加  2:删除 4: 更新
 	PermFlag int `db:"perm_flag"`
 }
